requirements_ui.txt
gradio>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
colorama>=0.4.6
urllib3>=2.0.0 